Update upload.sh
Changed remote directory for rsync to reflect changes in the structure of the server. The folder changed from webcs to www. This needs to be done to all upload.sh scripts. I'll test on this repository and then change the others.

ps.: Authorised keys also need to be updated if they use full commands.