Made a Variable into a Member to allow expressions of the form
Motor.F in p-models.

I have tried to restrict this change to the physmod plugins, but could
not do it without duplicating a lot of code. Notice that this change
does not add much to the metamodel, only an interface to the variable
class that categorises it as a member.